Extract stub classes to their own file to make them ignorable by PHPStan
